BUG/MEDIUM: mux-fcgi: Don't add private connections in available connection list
When a stream is detached from a backend private connection, we must not insert
it in the available connection list. In addition, we must be sure to remove it
from this list. To ensure it is properly performed, this part has been slightly
refactored to clearly split processing of private connections from the others.

This patch should probably be backported to 2.2.

(cherry picked from commit 29ae7ffed9c375be0c6d9204a100a5e950252362)
Signed-off-by: Willy Tarreau <w@1wt.eu>