BUG/MINOR: tcp-rules: Preserve the right filter analyser on content eval abort
When a tcp-request or a tcp-response content ruleset evaluation is aborted, the
corresponding FLT_END analyser must be preserved, if any. But because of a typo
error, on the tcp-response content ruleset evaluation, we try to preserve the
request analyser instead of the response one.

This patch must be backported to 2.2.

(cherry picked from commit 19dbf2d625b11e23cd45ef9d9678809a9ad602b9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>