BUG/MINOR: lua: Abort execution of actions that yield on a final evaluation
A Lua action may yield. It may happen because the action returns explicitly
act.YIELD or because the script itself yield. In the first case, we must abort
the script execution if it is the final rule evaluation, i.e if the
ACT_OPT_FINAL flag is set. The second case is already covered.

This patch must be backported to 2.2.

(cherry picked from commit 498c483009bfb69e75a35403c758b832d4d997d3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>