BUG/MEDIUM: mux-h2: Don't fail if nothing is parsed for a legacy chunk response
The parsing of incomplete chunk formatting in legacy HTTP mode was fixed by the
commit 6ad7cd981 ("BUG/MEDIUM: mux-h2: Emit an error if the response chunk
formatting is incomplete"). But a bug was introduced. When nothing is parsed, an
error is also returned. It may happens if the parsing is stopped just after a
chunk CRLF. When we try to parse the following chunk size, if there is no more
data to parse, an protocol error is returned. It is obviously wrong.

The patch was directly introduced on 2.0, there is no upstream commit ID for
this patch.

This patch is related to issue #798 and reported in comment in issue #768. It
must be backported to 1.8.

(cherry picked from commit 307f31ec34926965ef5c8f4f5b1d7840538f8246)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>